# policy: improve CNP initial sync