Check for the existence of CNP egress rules before translating
since we only do translation for egress rules.
This reduces the time cost of an add event of CNP without egress rules
and makes cilium agent start faster in a cluster with large amount of
CNPs and services.

Time cost comparison in a cluster with 10k CCNPs (all without egress rule), 44k services/endpoints:

Before:

    per CNP add event time cost              20 ~ 30 ms
    overall time cost of initial CNP sync    255 s

After:

    per CNP add event time cost              0.5 ~ 1 ms
    overall time cost of initial CNP sync    15 s

Signed-off-by: Jaff Cheng <jaff.cheng.sh@gmail.com>