@christarazi 2021-03-30
> > rocket Did you happen to make any time comparisons before and after this commit?
> 
> Yes, in one of our clusters with 10k CCNPs (all without egress rule), 44k services/endpoints, time cost of each CNP add event reduced from 20 ~ 30ms to 0.5 ~ 1ms, and the overall time cost of initial CNP sync reduced from 4m15s (exceeds k8s-sync-timeout) to 15s.

Awesome, could you encode that in the commit msg?